create main.py and update masks.py
