+++ pyproject.toml
@@ -15,6 +15,11 @@ black = "^24.4.2"
 mypy = "^1.10.1"
 isort = "^5.13.2"
 
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.2.2"
+pytest-cov = "^5.0.0"
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
