+++ widget.py
@@ -1,4 +1,4 @@
-from project_2.src.masks import get_mask_account, get_mask_card_number
+from src import get_mask_account, get_mask_card_number
 
 
 def mask_account_card(card_number: str) -> str:
